Revises Methods For v1 DAPI
This commit makes numerous revisions to the underlying datastructure and methods of the api client. These are breaking changes designed to bring the api client to v1 with the release of FMS 17’s v1 Data API.

It looks like the FMS 16’s Data API will be stopped on September 27th 2018. Since FMS 16 DAPI is being stopped I do not find it worthwhile to keep backwards compatibility.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>