1.2.0 Release
This commit represents the 1.20 release. This release contains an upgrade to the create method. You can now add the option to merge the record data with the record id and mod Id upon successful creation. This makes it easier to chain create then create command alongisde other methods without losing access to the original data.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>