1.6.0 Release
This commit signifies the 1.6.0 release. This release adds the client method of login and destroy.

The login method will start a filemaker session and return an object with a token property.

The destroy method will close a filemaker session and remove the client from the datastore.

This release contains breaking changes:

The client script method has switch the input of layout and script parameters. The script method now matches all other non authentication methods by requiring a layout as the first parameter.

The client fieldId and recordId utility methods have been removed from the client class. These utility methods are available alongside the transform method on the project export.

Signed-off-by: Lui de la Parra <Lui@mutesymphony.com>