1.7.0 Release
This commit signifies the 1.7.0 release.

This release contains breaking changes.

This release updates the client upload method to accept either a path to a file or a buffer.  This release also revises the client upload method to replace the field repetition with a parameters object. Set parameters.fieldRepetition to set a container field rep.

This release also adds client and request proxy, timeout, and agent configurations. Axios request parameters can now be set either when creating the client or when making the request.

Signed-off-by: Lui de la Parra <lui@mutesymphony.com>