Revises Scripting Methods
This commit revises client methods dealing with scripts. The client now has a run methoda and a script method.

The script method will trigger a single script.

The run method will fire an array of scripts. The array of scripts accepts a  maximum of three scripts corresponding with script execution phases.