2.1.0 Release
This commit represents the 2.1.0 release. This release contains a minor change to the agent watcher function. The watcher function is now held on the global scope and cleared when no longer in use.

The globalization of the watcher should prevent any possible memory reference collisions between clients and allow more effiecent use of the watcher method.