Fixes Session Start
This commit fixes a bug found via node-red-contrib-filemaker. This bug would not allow a token to be refreshed after initial use.

This commit revises the agent watch method to use a recursive setTimeout instead of SetInterval and delay.

The default delay has been incremented from a crazy 1 to a more sane 800. I don’t want to set the delay much higher for risk of adding an artifical delay.

This commit also adds a timeout to the request to FileMaker to open a session.

Finally, this commit revises the connection start method to filter expired tokens.

luidog/node-red-contrib-filemaker#26