Revises Agent Model
This commit revises the agent model request and watch methods. This commit should solve issue #71 and causes newly added tests to pass.

This commit revises the request method to eject any current intercepter when a new request is called.

This commit also revises the agent watch method to only create a session when their are pending requests and wait until there is an available session before queuing requests.

Signed-off-by: Lui de la Parra <Lui@mutesymphony.com>